use plant context

+++ app/app/(tabs)/index.tsx
@@ -4,57 +4,68 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import { Flower2 } from 'lucide-react-native';
 import * as Progress from 'react-native-progress';
 import { bleService } from '../../ble/BLEService';
+import { usePlants } from '../context/PlantsContext';
 
 
 export default function HomeScreen() {
-  const [moisture, setMoisture] = useState(0);
-  const [connected, setConnected] = useState(false);
+  const { selectedPlant, updateMoisture } = usePlants();
+
+  const moisture = selectedPlant?.moisture ?? 0;
+  // const [connected, setConnected] = useState(false);
 
   useEffect(() => {
+    if(!selectedPlant) return;
+
     bleService.scanForDevice(async (device) => {
       await bleService.connect(device);
-      setConnected(true);
 
       bleService.subscribeToMoisture((percent) => {
-        setMoisture(percent);
+        updateMoisture(selectedPlant.id, percent);
       });
     });
 
     return () => bleService.destroy();
-  }, []);
+  }, [selectedPlant, updateMoisture, selectedPlant?.id]);
+
+  if(!selectedPlant){
+    return (
+      <SafeAreaView style={styles.container}>
+        <Text style={{ color: 'red' }}>No plant selected</Text>
+      </SafeAreaView>
+    );
+  }
   
 
   return (
     <SafeAreaView style={styles.container}>
-      {!connected ? (
-        <Text>Connecting...</Text> // @todo: probably loading spinner
-      ) : (
-        <>
-          <Flower2 size={150}/>
-           <Progress.Bar 
-            progress={moisture / 100} 
-            width={200} 
-            color={ moisture < 30 ? 'red' : moisture > 50 ? '#019CE0' : 'orange'}
-          />
-          {(moisture) < 30 && (
-            <Text style={{ color: 'red', fontSize: 20, fontWeight: 'bold' }}>
-              Water soon
-            </Text>
-          )}
-          <Text style={{ fontSize: 24, fontWeight: 'bold', marginTop: 30 }}>
-            Moisture: {moisture}% 
-          </Text>
+      <Text style={{ fontSize: 28, fontWeight: 'bold', marginBottom: 20 }}>
+        {selectedPlant.name}
+      </Text>
 
+      <Flower2 size={150}/>
 
-          {/** @todo: not sure if we need to keep this? maybe at least move somewhere else */}
-          <TouchableOpacity 
-            style={styles.button} 
-            onPress={() => bleService.writeLED(moisture < 30 ? "1" : "0")}
-          >
-            <Text style={styles.buttonText}>Toggle LED</Text>
-          </TouchableOpacity>
-        </>
+      <Progress.Bar 
+        progress={moisture / 100} 
+        width={200} 
+        color={ moisture < 30 ? 'red' : moisture > 50 ? '#019CE0' : 'orange'}
+      />
+      {(moisture) < 30 && (
+        <Text style={{ color: 'red', fontSize: 20, fontWeight: 'bold' }}>
+          Water soon
+        </Text>
       )}
+      <Text style={{ fontSize: 24, fontWeight: 'bold', marginTop: 30 }}>
+        Moisture: {moisture}% 
+      </Text>
+
+
+      {/** @todo: not sure if we need to keep this? maybe at least move somewhere else */}
+      <TouchableOpacity 
+        style={styles.button} 
+        onPress={() => bleService.writeLED(moisture < 30 ? "1" : "0")}
+      >
+        <Text style={styles.buttonText}>Toggle LED</Text>
+      </TouchableOpacity>
     </SafeAreaView>
   );
 }
